made some slight changes to the CheckListBox/CheckComboBox.
IMPORTANT: There is a SelectedItems property for both controls, but it is intended to be ReadOnly.  You cannot data bind to ReadOnly DependencyProperties, so in order to enable the scenario of data binding to the SelectedItems from a ViewModel, I had to leave it as a public get/set property. For this binding to work correctly, you MUST set the binding Mode=OneWayToSource for the SelectedItems property.