Prevent crash from setting negative size
This change temporarily fixes several issues, as described in #1381 and #1379. The fix is such that the AvalonDock will no longer crash when the control goes out of bounds. However, this does not prevent the actual control from going out of bounds. Therefore, this fix should only be used as a stop-gap until an actual fix can be implemented.