Fixed LayoutRoot implementation of IXmlSerializable.ReadXml to read the end element.
Implementations of IXmlSerializable.ReadXml are required to read the entire element from beginning to end, including all of its contents. Prior to this change it would not read the end element from the reader. This would cause issues if the LayoutRoot was being serialized as part of a parent object's xml serialization as the XmlReader would be left on an unexpected element.