Drop the confusing wording mentioning "Open Source"
A license that limits commercial use is by definition not open source.
It for example violates section 6 of https://opensource.org/osd.

Dropping the confusing and deceptive wording here simplifies the
understanding of the license restrictions.

Also the image
https://user-images.githubusercontent.com/29377763/69274195-d9382200-0ba7-11ea-9ab7-bfce3126f35a.png
should be replaced by something not falsely stating that it is an open
source license.